Move BuildStepType
