+++ jib-core/src/main/java/com/google/cloud/tools/jib/api/event/BuildStepType.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package com.google.cloud.tools.jib.api.event.events;
+package com.google.cloud.tools.jib.api.event;
 
 /** Types corresponding to steps in the containerization process. */
 public enum BuildStepType {
