+++ jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/BuildAndCacheApplicationLayerStep.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.jib.builder.steps;
 
-import com.google.cloud.tools.jib.api.event.events.BuildStepType;
+import com.google.cloud.tools.jib.api.event.BuildStepType;
 import com.google.cloud.tools.jib.api.event.events.LayerCountEvent;
 import com.google.cloud.tools.jib.api.event.events.LogEvent;
 import com.google.cloud.tools.jib.async.AsyncStep;
