+++ jib-core/src/test/java/com/google/cloud/tools/jib/api/event/events/ProgressEventTest.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.jib.api.event.events;
 
 import com.google.cloud.tools.jib.api.event.Allocation;
+import com.google.cloud.tools.jib.api.event.BuildStepType;
 import com.google.cloud.tools.jib.api.event.EventHandlers;
 import com.google.cloud.tools.jib.api.event.JibEventType;
 import java.util.HashMap;
